Add initial setup for ArcaneArchivist Web API project
This commit introduces the boilerplate configuration of the ArcaneArchivist Web API project in the solution. It includes the base Web API project files, a Dockerfile to support containerized deployment, necessary .gitignore, and .dockerignore files, as well as default config files for logging and allowed hosts. The application endpoint at "/" is set to return "Hello World!".