Add MagicCards related entities and configure DB
Added new MagicCards related entities along with their configurations for the database. The previous ArcaneArchivist.WebApi project was deleted and re-created at a different location, with the same settings and configurations but with additional packages and settings related to Entity Framework. A migration was made to apply these configurations to the database.