Update GitHub workflow to convert test results to image
The code updates the GitHub workflow, changing it from merely extracting images from HTML to converting the entire HTML test results into an image. The new action utilizes the getsentry/action-html-to-image tool instead of the ericvergnaud/action-html-img-extractor. The changed process simplifies the representation of the test results and standardizes the output.